Move debug helpers into DebugInformation
